feat: move file signing to own util file
